Generator management
Two modes can now be used: 1 is were pods will be created for both the
generator and the swap pod, the other is a mode where the generators are
started upfront and keep running. When starting a test, we then only
create the swap pod, using the generator pods as described in the
rapid.pods file.
The rapid.pods file can now also specify for each pod,
- which pod yaml file should be used,
- which namespace or if a unique namespace needs to be created for this test,
- if nodeport service should be used,
- and the simd length