add irq score
A score is now calculated during the irq testing, so we have a single
figure as an outcome. 100% means that all pollmode iterations were done
very fast end were not interrupted by anything else (e.g. IRQs). That
means the system is very well tuned. When the system returns scores
below 80%, it is probably a good idea to inspect and find out why.
The score is calculated by subtracting a penalty with a heavier weigth
for longer interruptions